dom/media/mediasource/test/test_HaveMetadataUnbufferedSeek.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute before end of stream</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/webm");

  const arrayBuffer = await fetchWithXHR("seek.webm");
  sb.appendBuffer(new Uint8Array(arrayBuffer, 0, 67833));

  const target = 2;

  await once(v, "loadeddata");
  ok(v.readyState >= v.HAVE_CURRENT_DATA, "readyState is >= CURRENT_DATA");
  v.currentTime = target;

  await once(v, "seeking");
  is(v.readyState, v.HAVE_METADATA, "readyState is HAVE_METADATA");
  sb.appendBuffer(new Uint8Array(await fetchWithXHR("seek.webm"), 67833));
  await once(v, "seeked");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
